requirements.txt
ctransformers==0.2.5
sentence-transformers==2.2.2
pinecone-client==5.0.0
langchain==0.0.225
langchain-groq==0.1.6
Groq==0.9.0
flask==3.0.3
pypdf==4.3.0
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4
requests
tqdm==4.66.4
langchain==0.0.225
pinecone==5.0.0
-e .



